requirements.txt
streamlit
opencv-python
numpy
matplotlib
insightface
onnxruntime
pillow
